refactor: utilsディレクトリを削除
