+++ src/common/errors/index.ts
@@ -2,5 +2,6 @@ import AlreadyExistsError from './alreadyExistsError';
 import ClientError from './clientError';
 import NotFoundError from './notFoundError';
 import ServerError from './serverError';
+import getErrorMessage from './errorMessage';
 
-export { ClientError, AlreadyExistsError, NotFoundError, ServerError };
+export { ClientError, AlreadyExistsError, NotFoundError, ServerError, getErrorMessage };
